Se modifica estructura de las clases en las habitaciones La lectura de los 3 json se van a unificar en la clase GestionJSON Se agrega parametro en JSONUtiles para poder pasar en el metodo el nombre del json que se va a leer
