--- Habitaciones/src/ManejoJSON/GestionHabitacionesJSON.java
@@ -1,7 +1,0 @@
-package ManejoJSON;
-
-public class GestionHabitacionesJSON {
-
-    //MAPEO CONTEMPLANDO LAS CLASES POR CADA TIPO DE HABITACION
-
-}
