--- Habitaciones/src/Models/Habitaciones/Apartamento.java
@@ -1,7 +1,0 @@
-package Models.Habitaciones;
-
-import Models.Habitacion;
-
-public class Apartamento extends Habitacion {
-
-}
