--- Habitaciones/src/Models/Habitaciones/Doble.java
@@ -1,7 +1,0 @@
-package Models.Habitaciones;
-
-import Models.Habitacion;
-
-public class Doble extends Habitacion {
-
-}
