--- Habitaciones/src/Models/Habitaciones/Individual.java
@@ -1,7 +1,0 @@
-package Models.Habitaciones;
-
-import Models.Habitacion;
-
-public class Individual extends Habitacion {
-
-}
