--- Habitaciones/src/Models/Habitaciones/Presidencial.java
@@ -1,7 +1,0 @@
-package Models.Habitaciones;
-
-import Models.Habitacion;
-
-public class Presidencial extends Habitacion {
-
-}
