--- Habitaciones/src/Models/Habitaciones/Suite.java
@@ -1,7 +1,0 @@
-package Models.Habitaciones;
-
-import Models.Habitacion;
-
-public class Suite extends Habitacion {
-
-}
