+++ HotelManagement_I/src/main/java/controllers/crear/CrearReservaController.java
@@ -8,9 +8,7 @@ import javafx.scene.Node;
 import javafx.scene.control.*;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
-import models.Habitacion;
 import models.Pasajero;
-import models.Reserva;
 import models.Usuario;
 import services.GestionReservas;
