+++ HotelManagement_I/src/main/java/controllers/details/DetallesGeneralesController.java
@@ -8,7 +8,6 @@ import javafx.scene.control.Alert;
 import javafx.scene.control.Label;
 import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
-import models.Habitacion;
 import models.Pasajero;
 import models.Usuario;
