+++ HotelManagement_I/src/main/java/controllers/details/DetallesReservaController.java
@@ -8,7 +8,6 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Label;
 import javafx.stage.Stage;
-import models.Habitacion;
 import models.Pasajero;
 import models.Reserva;
 import models.Usuario;
