+++ HotelManagement_I/src/main/java/enums/EstadoHabitacion.java
@@ -1,4 +1,4 @@
-package Enums;
+package main.java.enums;
 
 public enum EstadoHabitacion {
     OCUPADA("Ocupada"),
