+++ HotelManagement_I/src/main/java/exceptions/CapacidadInvalidaException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package main.java.exceptions;
 
 public class CapacidadInvalidaException extends RuntimeException {
     public CapacidadInvalidaException(String message) {
