+++ HotelManagement_I/src/main/java/exceptions/NroHabitacionInvalidoException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package main.java.exceptions;
 
 public class NroHabitacionInvalidoException extends RuntimeException {
     public NroHabitacionInvalidoException(String message) {
