+++ HotelManagement_I/src/main/java/manejoJson/GestionJSON.java
@@ -1,0 +1,8 @@
+package main.java.manejoJson;
+
+//Se realiza el mapeo de los 3 Json del proyecto
+public class GestionJSON {
+
+    //MAPEO CONTEMPLANDO LAS CLASES POR CADA TIPO DE HABITACION
+
+}
