+++ HotelManagement_I/src/main/java/models/Habitacion/Doble.java
@@ -1,0 +1,5 @@
+package main.java.models.Habitacion;
+
+public class Doble extends Habitacion {
+
+}
