+++ HotelManagement_I/src/main/java/models/Habitacion/Suite.java
@@ -1,0 +1,5 @@
+package main.java.models.Habitacion;
+
+public class Suite extends Habitacion {
+
+}
