+++ HotelManagement_I/src/main/java/models/Pasajero.java
@@ -1,4 +1,4 @@
-package models;
+package main.java.models;
 
 public class Pasajero {
     private String nombre;
