+++ HotelManagement_I/src/main/java/models/Reserva.java
@@ -1,5 +1,4 @@
 package models;
-import models.Habitacion;
 
 
 import java.time.LocalDate;
