+++ HotelManagement_I/src/main/java/services/GestionHabitaciones.java
@@ -1,4 +1,6 @@
-package Models;
+package main.java.services;
+
+import main.java.models.Habitacion.Habitacion;
 
 import java.util.*;
