+++ HotelManagement_I/src/main/java/services/GestionarHabitaciones.java
@@ -3,7 +3,6 @@ package services;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
-import models.Habitacion;
 import java.util.ArrayList;
 import java.util.List;
